Fixed carrot overflow issue
A u8 is too small for the number of carrots, because a hare can have more than 255 carrots

+++ rust/simple-client/src/parsers/parse_memento.rs
@@ -40,7 +40,7 @@ pub fn parse_memento(message: &String, game_data: &mut GameData) {
                         let mut team: Option<Team> = None;
                         let mut position: Option<u8> = None;
                         let mut salads: Option<u8> = None;
-                        let mut carrots: Option<u8> = None;
+                        let mut carrots: Option<u16> = None;
 
                         // Retreive the team
                         if let Some(attr) = e.attributes().find(|a| a.as_ref().unwrap().key == QName(b"team")) {
